requirements.txt
streamlit
psycopg2-binary
torch
faker
pandas
matplotlib
python-dotenv
